requirements.txt
boto3==1.34.84
botocore==1.34.84
python-dotenv==1.0.0
streamlit==1.30.0
langchain==0.1.16
langchain-community==0.0.32
giskard[llm]
